# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Operations Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];

export const TaskStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export type TaskPriority = (typeof TaskPriority)[keyof typeof TaskPriority];

export const TaskPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface Task {
  id: string;
  title: string;
  description?: string;
  status: TaskStatus;
  priority: TaskPriority;
  dueDate?: string | null;
  assignee?: string | null;
  tags: string[];
  createdAt: string;
  updatedAt: string;
  isOverdue: boolean;
}

export interface TaskListResponse {
  tasks: Task[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export type CreateTaskBodyStatus =
  (typeof CreateTaskBodyStatus)[keyof typeof CreateTaskBodyStatus];

export const CreateTaskBodyStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export type CreateTaskBodyPriority =
  (typeof CreateTaskBodyPriority)[keyof typeof CreateTaskBodyPriority];

export const CreateTaskBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface CreateTaskBody {
  title: string;
  description?: string;
  status?: CreateTaskBodyStatus;
  priority: CreateTaskBodyPriority;
  dueDate?: string | null;
  assignee?: string | null;
  tags?: string[];
}

export type UpdateTaskBodyStatus =
  (typeof UpdateTaskBodyStatus)[keyof typeof UpdateTaskBodyStatus];

export const UpdateTaskBodyStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export type UpdateTaskBodyPriority =
  (typeof UpdateTaskBodyPriority)[keyof typeof UpdateTaskBodyPriority];

export const UpdateTaskBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface UpdateTaskBody {
  title?: string;
  description?: string;
  status?: UpdateTaskBodyStatus;
  priority?: UpdateTaskBodyPriority;
  dueDate?: string | null;
  assignee?: string | null;
  tags?: string[];
}

export interface AnalyticsSummary {
  total: number;
  pending: number;
  inProgress: number;
  completed: number;
  overdue: number;
  completionRate: number;
}

export type PriorityCountPriority =
  (typeof PriorityCountPriority)[keyof typeof PriorityCountPriority];

export const PriorityCountPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface PriorityCount {
  priority: PriorityCountPriority;
  count: number;
}

export type StatusCountStatus =
  (typeof StatusCountStatus)[keyof typeof StatusCountStatus];

export const StatusCountStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export interface StatusCount {
  status: StatusCountStatus;
  count: number;
}

export interface TrendPoint {
  date: string;
  completed: number;
  created: number;
}

export type ListTasksParams = {
  status?: ListTasksStatus;
  priority?: ListTasksPriority;
  search?: string;
  page?: number;
  limit?: number;
};

export type ListTasksStatus =
  (typeof ListTasksStatus)[keyof typeof ListTasksStatus];

export const ListTasksStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export type ListTasksPriority =
  (typeof ListTasksPriority)[keyof typeof ListTasksPriority];

export const ListTasksPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;