# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Operations Dashboard API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: tasks
    description: Task management operations
  - name: analytics
    description: Dashboard analytics operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks
      description: Returns a filtered, paginated list of tasks
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_progress, completed]
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskListResponse"

    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Task deleted
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /analytics/summary:
    get:
      operationId: getAnalyticsSummary
      tags: [analytics]
      summary: Get dashboard analytics summary
      description: Returns counts for total, pending, in-progress, completed, and overdue tasks
      responses:
        "200":
          description: Analytics summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsSummary"

  /analytics/by-priority:
    get:
      operationId: getTasksByPriority
      tags: [analytics]
      summary: Get task counts grouped by priority
      responses:
        "200":
          description: Task counts by priority
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PriorityCount"

  /analytics/by-status:
    get:
      operationId: getTasksByStatus
      tags: [analytics]
      summary: Get task counts grouped by status
      responses:
        "200":
          description: Task counts by status
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StatusCount"

  /analytics/completion-trend:
    get:
      operationId: getCompletionTrend
      tags: [analytics]
      summary: Get task completion trend over the last 14 days
      responses:
        "200":
          description: Completion trend data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrendPoint"

  /analytics/overdue:
    get:
      operationId: getOverdueTasks
      tags: [analytics]
      summary: Get all overdue tasks (not completed, past due date)
      responses:
        "200":
          description: List of overdue tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message

    Task:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed]
        priority:
          type: string
          enum: [low, medium, high, critical]
        dueDate:
          type: string
          format: date-time
          nullable: true
        assignee:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        isOverdue:
          type: boolean
      required:
        - id
        - title
        - status
        - priority
        - tags
        - createdAt
        - updatedAt
        - isOverdue

    TaskListResponse:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - tasks
        - total
        - page
        - limit
        - totalPages

    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed]
          default: pending
        priority:
          type: string
          enum: [low, medium, high, critical]
          default: medium
        dueDate:
          type: string
          format: date-time
          nullable: true
        assignee:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
      required:
        - title
        - priority

    UpdateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed]
        priority:
          type: string
          enum: [low, medium, high, critical]
        dueDate:
          type: string
          format: date-time
          nullable: true
        assignee:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string

    AnalyticsSummary:
      type: object
      properties:
        total:
          type: integer
        pending:
          type: integer
        inProgress:
          type: integer
        completed:
          type: integer
        overdue:
          type: integer
        completionRate:
          type: number
          format: float
      required:
        - total
        - pending
        - inProgress
        - completed
        - overdue
        - completionRate

    PriorityCount:
      type: object
      properties:
        priority:
          type: string
          enum: [low, medium, high, critical]
        count:
          type: integer
      required:
        - priority
        - count

    StatusCount:
      type: object
      properties:
        status:
          type: string
          enum: [pending, in_progress, completed]
        count:
          type: integer
      required:
        - status
        - count

    TrendPoint:
      type: object
      properties:
        date:
          type: string
        completed:
          type: integer
        created:
          type: integer
      required:
        - date
        - completed
        - created