# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Operations Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns a filtered, paginated list of tasks
 * @summary List all tasks
 */
export const listTasksQueryPageDefault = 1;
export const listTasksQueryLimitDefault = 20;

export const ListTasksQueryParams = zod.object({
  status: zod.enum(["pending", "in_progress", "completed"]).optional(),
  priority: zod.enum(["low", "medium", "high", "critical"]).optional(),
  search: zod.coerce.string().optional(),
  page: zod.coerce.number().default(listTasksQueryPageDefault),
  limit: zod.coerce.number().default(listTasksQueryLimitDefault),
});

export const ListTasksResponse = zod.object({
  tasks: zod.array(
    zod.object({
      id: zod.string(),
      title: zod.string(),
      description: zod.string().optional(),
      status: zod.enum(["pending", "in_progress", "completed"]),
      priority: zod.enum(["low", "medium", "high", "critical"]),
      dueDate: zod.coerce.date().nullish(),
      assignee: zod.string().nullish(),
      tags: zod.array(zod.string()),
      createdAt: zod.coerce.date(),
      updatedAt: zod.coerce.date(),
      isOverdue: zod.boolean(),
    }),
  ),
  total: zod.number(),
  page: zod.number(),
  limit: zod.number(),
  totalPages: zod.number(),
});

/**
 * @summary Create a new task
 */
export const createTaskBodyStatusDefault = `pending`;
export const createTaskBodyPriorityDefault = `medium`;

export const CreateTaskBody = zod.object({
  title: zod.string(),
  description: zod.string().optional(),
  status: zod
    .enum(["pending", "in_progress", "completed"])
    .default(createTaskBodyStatusDefault),
  priority: zod
    .enum(["low", "medium", "high", "critical"])
    .default(createTaskBodyPriorityDefault),
  dueDate: zod.coerce.date().nullish(),
  assignee: zod.string().nullish(),
  tags: zod.array(zod.string()).optional(),
});

/**
 * @summary Get a task by ID
 */
export const GetTaskParams = zod.object({
  id: zod.coerce.string(),
});

export const GetTaskResponse = zod.object({
  id: zod.string(),
  title: zod.string(),
  description: zod.string().optional(),
  status: zod.enum(["pending", "in_progress", "completed"]),
  priority: zod.enum(["low", "medium", "high", "critical"]),
  dueDate: zod.coerce.date().nullish(),
  assignee: zod.string().nullish(),
  tags: zod.array(zod.string()),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
  isOverdue: zod.boolean(),
});

/**
 * @summary Update a task
 */
export const UpdateTaskParams = zod.object({
  id: zod.coerce.string(),
});

export const UpdateTaskBody = zod.object({
  title: zod.string().optional(),
  description: zod.string().optional(),
  status: zod.enum(["pending", "in_progress", "completed"]).optional(),
  priority: zod.enum(["low", "medium", "high", "critical"]).optional(),
  dueDate: zod.coerce.date().nullish(),
  assignee: zod.string().nullish(),
  tags: zod.array(zod.string()).optional(),
});

export const UpdateTaskResponse = zod.object({
  id: zod.string(),
  title: zod.string(),
  description: zod.string().optional(),
  status: zod.enum(["pending", "in_progress", "completed"]),
  priority: zod.enum(["low", "medium", "high", "critical"]),
  dueDate: zod.coerce.date().nullish(),
  assignee: zod.string().nullish(),
  tags: zod.array(zod.string()),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
  isOverdue: zod.boolean(),
});

/**
 * @summary Delete a task
 */
export const DeleteTaskParams = zod.object({
  id: zod.coerce.string(),
});

/**
 * Returns counts for total, pending, in-progress, completed, and overdue tasks
 * @summary Get dashboard analytics summary
 */
export const GetAnalyticsSummaryResponse = zod.object({
  total: zod.number(),
  pending: zod.number(),
  inProgress: zod.number(),
  completed: zod.number(),
  overdue: zod.number(),
  completionRate: zod.number(),
});

/**
 * @summary Get task counts grouped by priority
 */
export const GetTasksByPriorityResponseItem = zod.object({
  priority: zod.enum(["low", "medium", "high", "critical"]),
  count: zod.number(),
});
export const GetTasksByPriorityResponse = zod.array(
  GetTasksByPriorityResponseItem,
);

/**
 * @summary Get task counts grouped by status
 */
export const GetTasksByStatusResponseItem = zod.object({
  status: zod.enum(["pending", "in_progress", "completed"]),
  count: zod.number(),
});
export const GetTasksByStatusResponse = zod.array(GetTasksByStatusResponseItem);

/**
 * @summary Get task completion trend over the last 14 days
 */
export const GetCompletionTrendResponseItem = zod.object({
  date: zod.string(),
  completed: zod.number(),
  created: zod.number(),
});
export const GetCompletionTrendResponse = zod.array(
  GetCompletionTrendResponseItem,
);

/**
 * @summary Get all overdue tasks (not completed, past due date)
 */
export const GetOverdueTasksResponseItem = zod.object({
  id: zod.string(),
  title: zod.string(),
  description: zod.string().optional(),
  status: zod.enum(["pending", "in_progress", "completed"]),
  priority: zod.enum(["low", "medium", "high", "critical"]),
  dueDate: zod.coerce.date().nullish(),
  assignee: zod.string().nullish(),
  tags: zod.array(zod.string()),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
  isOverdue: zod.boolean(),
});
export const GetOverdueTasksResponse = zod.array(GetOverdueTasksResponseItem);